Add `update-notifier` (#9)
Alternative to #8 with a module that pulls in more dependencies, but which also have more features and eg. defers printing the result until the script quits, disables itself on CI, always prints to stderr etc

Has quite a few more users as well. A good first step to try this out probably.